requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
--trusted-host pypi.tuna.tsinghua.edu.cn

PyQt6==6.10.2
PyQt6-Fluent-Widgets==1.11.1
mss>=9.0.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pywin32>=306
pyvda==0.5.0
Pillow>=10.0.0
loguru>=0.7.0
onepush==1.9.0
pydantic>=2.5.0
PyYAML>=6.0.0
opencv-python>=4.9.0
numpy>=1.26.0
keyboard>=0.13.5
rapidocr==3.8.1
onnxruntime==1.19.2
winotify==1.1.0
